requirements.txt
PyJWT==2.12.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
